pyproject.toml
[tool.poetry]
name = "djangoweber"
version = "0.1.0"
description = ""
authors = ["Diego Alonso Ojeda Medina <daom89@gmail.com>"]
readme = "README.md"
packages = [{include = "django_weber"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
mysqlclient = "^2.2.0"
django-safedelete = "^1.3.2"
django-dump-load-utf8 = "^0.0.4"
django-environ = "^0.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
